+++ src/Router.jsx
@@ -4,11 +4,11 @@ import Education from "./pages/Education";
 import Home from "./pages/Home";
 import Intro from "./pages/Intro";
 import MyPage from "./pages/MyPage";
-import Register from "./pages/Register";
+import SignUp from "./pages/SignUp";
 import MemberListPage from "./pages/MemberListPage";
 import Login from "./pages/Login";
 import GithubLoginPage from "./pages/GithubLoginPage";
-import Certification from "./components/register/Certification";
+import Certification from "./components/signup/Certification";
 import { getCookieToken } from "./cookie/cookie";
 
 export default function Router() {
@@ -24,7 +24,7 @@ export default function Router() {
           path={"/login"}
           element={isLogin ? <Navigate to="/home" /> : <Login />}
         />
-        <Route path={"/register"} element={<Register />} />
+        <Route path={"/register"} element={<SignUp />} />
         <Route path={"/home"} element={<Home />} />
         <Route path={"/education"} element={<Education />} />
         <Route path={"/mypage"} element={<MyPage />} />
